v2.0.1

+++ app/src/main/java/net/muxi/huashiapp/common/data/CardSumData.java
@@ -6,9 +6,9 @@ package net.muxi.huashiapp.common.data;
 
 public class CardSumData {
     public String time;
-    public float sum;
+    public double sum;
 
-    public CardSumData(String time, float sum) {
+    public CardSumData(String time, double sum) {
         this.time = time;
         this.sum = sum;
     }
